* uia2atk/UIAutomationWinforms/UIAutomationWinforms/Mono.UIAutomation.Winforms.Behaviors/Button/InvokeProviderBehavior.cs:   Whitespace fix.
* uia2atk/UIAutomationWinforms/UIAutomationWinforms/Mono.UIAutomation.Winforms.Behaviors/TextBox/ValueProviderBehavior.cs:
  Use BeginInvoke to set value, when necessary.

* uia2atk/UIAutomationWinforms/UIAutomationWinforms/Mono.UIAutomation.Winforms.Behaviors/UpDownBase/RangeValueProviderBehavior.cs:
  Removing extraneous cast.

svn path=/trunk/uia2atk/; revision=119504

+++ UIAutomationWinforms/UIAutomationWinforms/Mono.UIAutomation.Winforms.Behaviors/UpDownBase/RangeValueProviderBehavior.cs
@@ -148,7 +148,7 @@ namespace Mono.UIAutomation.Winforms.Behaviors.UpDownBase
 				                           new object [] { value });
 				return;
 			}
-			numericUpDown.Value = (decimal) value;
+			numericUpDown.Value = value;
 		}
 			
 		#endregion
